lambda/auth/handler.py
import json
import os
import time
import jwt
from datetime import datetime, timedelta

def authenticate(event, context):
    """Placeholder auth Lambda - validates signature and returns JWT"""
    try:
        body = json.loads(event.get('body', '{}'))
        address = body.get('address')
        
        if not address:
            return {
                'statusCode': 400,
                'body': json.dumps({'error': 'Missing wallet address'})
            }
        
        # TODO: Verify wallet signature
        # TODO: Check NFT ownership
        
        # For now, generate JWT for any address
        secret = os.environ['JWT_SECRET_NAME']  # Will get from Secrets Manager
        token = jwt.encode({
            'address': address,
            'exp': datetime.utcnow() + timedelta(hours=24)
        }, 'temp-secret', algorithm='HS256')
        
        return {
            'statusCode': 200,
            'body': json.dumps({'token': token})
        }
    except Exception as e:
        return {
            'statusCode': 500,
            'body': json.dumps({'error': str(e)})
        }