lambda/auth/requirements.txt
PyJWT==2.8.0
eth-account==0.10.0
web3==6.11.3